Update jwt converters
- Create abstract converter using method template
- Delegate provider converter logic to child classes though impl of the method providerAuthorities
- Add utility class ConverterUtil to build authorities naming
