+++ rest/src/main/java/org/wcdevs/blog/core/rest/converter/ConverterUtil.java
@@ -1,0 +1,10 @@
+package org.wcdevs.blog.core.rest.converter;
+
+class ConverterUtil {
+  private ConverterUtil() {
+  }
+
+  static String toAuthRoleName(String name) {
+    return "ROLE_" + name;
+  }
+}
